requirements.txt
fastapi
uvicorn
SQLAlchemy
mysql-connector-python
python-dotenv
jinja2